Consolidates schema generator logic to FormulateSchema
