+++ src/libs/utils.js
@@ -303,22 +303,3 @@ export function isEmpty (value) {
     )
   )
 }
-
-/**
- * Given an object and an index, complete an object for schema-generation.
- * @param {object} item
- * @param {int} index
- */
-export function createItem (item, index) {
-  if (item && typeof item === 'object' && !Array.isArray(item)) {
-    const { children = null, component = 'FormulateInput', depth = 1, ...attrs } = item
-    const type = component === 'FormulateInput' ? (attrs.type || 'text') : ''
-    const name = attrs.name || type || 'el'
-    const key = attrs.id || `${name}-${depth}-${index}`
-    const els = Array.isArray(children)
-      ? children.map(child => Object.assign(child, { depth: depth + 1 }))
-      : false
-    return Object.assign({ key, depth, attrs, component }, els ? { children: els } : {})
-  }
-  return null
-}
